+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the "wipeout" extension will be documented in this file.
 
 ## [Unreleased]
 
+## [0.0.10]
+
  - Added a context menu to the SceneGraph to jump to textures
  - Added .fnt loader for Wipeout HD
  - Improve VEXX materials caching